Fix generateStaticParams error with proper Server/Client Component architecture
- Split projects/[id] page into Server Component + ProjectDetailClient
- Split blog/[id] page into Server Component + BlogPostClient
- Added generateStaticParams() functions for static export compatibility
- Server Components handle data fetching and static generation
- Client Components handle interactive features (framer-motion, ScrollReveal)
- All dynamic routes now work correctly in development mode
- Fixed the 'generateStaticParams() required with output: export' error

Routes working:
- /projects/1, /projects/2, /projects/3 ✅
- /blog/1, /blog/2, /blog/3 ✅
- Proper error handling for non-existent routes ✅
